1.add default settings for rg35xx-sp in drastic_steward 2.fix the problem that custom setting values are not saved. 3.fix the problem that does not show in the last line in the extra controls menu 4.fix to run in np2kai even if .bmp font does not exist 5.fix to run dpad-toggle when func+select is entered at rg35xx-sp, rg28xx and rg35xx-p only 6.add gmu 와 sdl-flip-clock packages when clean building 7.Fix a problem that does not work when entering custom settings 8.Modify the setting value path for openmsx to the default position

+++ package/batocera/core/batocera-configgen/configgen/configgen/generators/libretro/libretroGenerator.py
@@ -287,7 +287,8 @@ class LibretroGenerator(Generator):
             commandArray = [batoceraFiles.batoceraBins[system.config['emulator']], "-L", retroarchCore, "--config", system.config['configfile']]
         elif system.name == 'pc98':
             romPC98Name, romExtension = os.path.splitext(romName)
-            shutil.copyfile(os.path.dirname(os.path.abspath(rom)) + "/" + romPC98Name + ".bmp", "/userdata/bios/np2kai/font.bmp")
+            if os.path.isfile(os.path.dirname(os.path.abspath(rom)) + "/" + romPC98Name + ".bmp"):
+                shutil.copyfile(os.path.dirname(os.path.abspath(rom)) + "/" + romPC98Name + ".bmp", "/userdata/bios/np2kai/font.bmp")
             retroarchCore = batoceraFiles.retroarchCores + system.config['core'] + "_libretro.so"
             commandArray = [batoceraFiles.batoceraBins[system.config['emulator']], "-L", retroarchCore, "--config", system.config['configfile']]
         else:
